refactor: ColorPicker 컴포넌트 타입스크립트로 전환
